# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "York academy",
        "description": "York academy",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/"
        },
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/course_ads/search": {
            "get": {
                "tags": [
                    "Course Ads"
                ],
                "summary": "Search for advanced course ads",
                "description": "Search course ads with advanced filters",
                "operationId": "searchAdvanceCourseAds",
                "parameters": [
                    {
                        "name": "languages[]",
                        "in": "query",
                        "description": "Filter by languages",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            },
                            "example": [
                                "English"
                            ]
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Filter by course code",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "522828"
                        }
                    },
                    {
                        "name": "venue_ids[]",
                        "in": "query",
                        "description": "Filter by venue IDs",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            },
                            "example": [
                                29
                            ]
                        }
                    },
                    {
                        "name": "category_ids[]",
                        "in": "query",
                        "description": "Filter by category IDs",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            },
                            "example": [
                                19
                            ]
                        }
                    },
                    {
                        "name": "months[]",
                        "in": "query",
                        "description": "Filter by months",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "years[]",
                        "in": "query",
                        "description": "Filter by years",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully"
                    },
                    "400": {
                        "description": "Invalid request data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/course_ads/{id}": {
            "get": {
                "tags": [
                    "Course Ads"
                ],
                "summary": "Get course ad by ID",
                "description": "Retrieve details of a specific course ad by its ID",
                "operationId": "getCourseAdsById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course ad",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course ad details retrieved successfully"
                    },
                    "404": {
                        "description": "Course ad not found"
                    }
                }
            }
        },
...
        "/frequently_questions/{id}": {
            "get": {
                "tags": [
                    "Frequently Questions"
                ],
                "summary": "Retrieve an FAQ",
                "description": "Get details of a specific FAQ by ID",
                "operationId": "543a3da1431e104137e5ea59d97718c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the FAQ to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "FAQ retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FrequentlyQuestion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "FAQ not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FrequentlyQuestion": {
                "title": "Frequently Asked Question",
                "description": "A model representing a frequently asked question and its answer",
                "properties": {
                    "id": {
                        "description": "ID of the FAQ",
                        "type": "integer"
                    },
                    "question": {
                        "description": "The question text",
                        "type": "string"
                    },
                    "answer": {
                        "description": "The answer text",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when the FAQ was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Timestamp when the FAQ was last updated",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}